Move state_with_localhost_with_port fixture
We will need it on other test modules that are upcoming.
